+++ vss_cli/plugins/compute_plugins/folder.py
@@ -42,7 +42,7 @@ def compute_folder_ls(ctx: Configuration, filter_by, show_all, sort, page):
     """
     params = dict(expand=1, sort='path,asc')
     if all(filter_by):
-        params['filter'] = ','.join(process_filters(filter_by))
+        params['filter'] = ';'.join(process_filters(filter_by))
     if all(sort):
         params['sort'] = f'{sort[0]},{sort[1]}'
     # get objects