+++ vss_cli/plugins/compute_plugins/image.py
@@ -5,7 +5,7 @@ import click
 from vss_cli import const, rel_opts as so
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
-from vss_cli.helper import format_output
+from vss_cli.helper import format_output, process_filters
 from vss_cli.plugins.compute import cli
 
 _LOGGING = logging.getLogger(__name__)
@@ -41,7 +41,7 @@ def compute_image_public_ls(
     """
     params = dict(expand=1, sort='name,asc')
     if all(filter_by):
-        params['filter'] = f'{filter_by[0]},{filter_by[1]}'
+        params['filter'] = ';'.join(process_filters(filter_by))
     if all(sort):
         params['sort'] = f'{sort[0]},{sort[1]}'
     # get objects