+++ vss_cli/plugins/upgrade.py
@@ -3,6 +3,7 @@ import logging
 from subprocess import call
 
 import click
+
 from vss_cli import vssconst
 from vss_cli.cli import pass_context
 from vss_cli.config import Configuration
@@ -14,18 +15,15 @@ _LOGGING = logging.getLogger(__name__)
 @click.group(
     'upgrade',
     invoke_without_command=True,
-    short_help='Upgrade VSS CLI and dependencies.')
+    short_help='Upgrade VSS CLI and dependencies.',
+)
 @click.argument(
     'upstream',
-    type=click.Choice(
-        ['stable', 'develop', 'branch']
-    ),
-    default='stable'
+    type=click.Choice(['stable', 'develop', 'branch']),
+    default='stable',
 )
 @click.option(
-    '--git-branch', '-b',
-    help='GitLab repository branch',
-    required=False
+    '--git-branch', '-b', help='GitLab repository branch', required=False
 )
 @pass_context
 def cli(ctx: Configuration, upstream, git_branch):
@@ -45,23 +43,21 @@ def cli(ctx: Configuration, upstream, git_branch):
 
     """
     lookup = {
-        'stable': {
-            'pkg': 'vss-cli',
-            'args': ['--upgrade']
-        },
+        'stable': {'pkg': 'vss-cli', 'args': ['--upgrade']},
         'develop': {
             'pkg': 'vss-cli',
             'args': [
                 '--upgrade',
                 '--index-url',
-                'https://test.pypi.org/legacy/'
-            ]
+                'https://test.pypi.org/legacy/',
+            ],
         },
         'branch': {
-            'pkg': f'git+https://gitlab-ee.eis.utoronto.ca'
-            f'/vss/vss-cli.git@{git_branch}',
-            'args': ['--upgrade']
-        }
+            'pkg': f'https://gitlab-ee.eis.utoronto.ca'
+            f'/vss/vss-cli/-/archive/'
+            f'{git_branch}/vss-cli-{git_branch}.zip',
+            'args': ['--upgrade'],
+        },
     }
     try:
         cmd_lookup = lookup[upstream]
@@ -69,35 +65,28 @@ def cli(ctx: Configuration, upstream, git_branch):
             if not git_branch:
                 raise click.BadOptionUsage(
                     'branch',
-                    'Missing --branch/-b option. '
-                    'Try master or develop.'
+                    'Missing --branch/-b option. ' 'Try master or develop.',
                 )
     except ValueError as ex:
         _LOGGING.error(f'Invalid upstream {upstream}: {ex}')
-        raise click.BadArgumentUsage(
-            f'Invalid upstream {upstream}'
-        )
+        raise click.BadArgumentUsage(f'Invalid upstream {upstream}')
     # assemble command
     cmd_args_str = ' '.join(cmd_lookup['args'])
-    cmd_bin = ctx.get_pip_binary()
+    cmd_bin = ctx.get_python_binary()
     # execute command
-    cmd_str = f"{cmd_bin} install {cmd_args_str} {cmd_lookup['pkg']}"
+    cmd_str = f"{cmd_bin} -m pip install {cmd_args_str} {cmd_lookup['pkg']}"
     _LOGGING.debug(f'Executing {cmd_str}')
     # calling command
-    exit_code = call(
-        cmd_str,
-        shell=True)
+    exit_code = call(cmd_str, shell=True)
     if exit_code > 0:
         raise click.ClickException(
-            f'Could not perform upgrade, please try: '
-            f'\n\t{cmd_str}')
-    else:
-        _LOGGING.debug(
-            f'Successfully executed upgrade command: {cmd_str}'
+            f'Could not perform upgrade, please try: ' f'\n\t{cmd_str}'
         )
+    else:
+        _LOGGING.debug(f'Successfully executed upgrade command: {cmd_str}')
     # all done
     ctx.secho(
         f'Successfully executed upgrade command '
         f'{EMOJI_UNICODE.get(":white_heavy_check_mark:")}',
-        fg='green'
+        fg='green',
     )