+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog 📝
 
+## [v0.5.1](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v0.5.1) (2020-02-14)
+
+**Bug Fixes:**
+
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute vm ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute floppy ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute domain ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute folder ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute image personal ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute image public ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute iso personal ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute iso public ls``: ``--filter`` does not support multiple instances.
+- [#240](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/240): ``compute net ls``: ``--filter`` does not support multiple instances.
+- [#241](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/241): ``upgrade``: upgrade command to prioritize to ``python3 -m pip``.
+- [#242](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/242): ``upgrade``:  error when ``git`` is not installed.
+
+
 ## [v0.5.0](https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v0.5.0) (2020-02-06)
 
 **Improvements:**